Run the modularized flexdashboard

Run the modularized flexdashboard

explore_data(dataset, data_dictionary = NULL, outcome_var = NULL)

Arguments

dataset

a data frame

data_dictionary

Optional. A data.frame describing covariates. This is passed on to DT:data.table() so it can be easily searched. There is no required format, but the example uses two columns: VariableName and Description.

explore_data gives you a simple shiny app to explore data, The app is tailored to the covariate types of the data.frame/data.table that you pass to it.

Each menu item covers a different part of exploratory data analysis with simple visualizations for discussion.

The tabs are automatically updated with variables depending on the variable types.

outcome_var

a categorical variable that describes an outcome to examine

Value

A Shiny App object that can either be run in the console or on an external shiny server such as shinyapps.io. For building the actual app.R, see build_shiny_app

Examples

library(ggplot2)
data(diamonds)
if(interactive()){
   burro::explore_data(diamonds, outcome_var="cut")
}

# example with data dictionary
# look at NHANES data (you need to have NHANES package installed)
library(NHANES)
data(NHANES)
if(interactive()){
  #this data dictionary is provided since NHANES doesn't have one
  data_dict <- read.csv(system.file("nhanes/data_dictionary.csv", package="burro"))
  burro::explore_data(NHANES[1:10,], data_dictionary=data_dict)
}

data(mtcars)
mtcars$cyl <- factor(mtcars$cyl)
mtcars$am <- factor(mtcars$am)
mtcars$gear <- factor(mtcar$gear)

#> Error in factor(mtcar$gear): object 'mtcar' not found

if(interactive()){
   burro::explore_data(mtcars)
}