Explore a dataaset using a simple app

Explore a dataaset using a simple app

explore_data(dataset, covariates = NULL, outcome_var,
  data_dictionary = NULL)

Arguments

dataset

a data frame

covariates

Optional. a character vector of column names you want to include in the dataset

outcome_var

a categorical variable that describes an outcome to examine

data_dictionary

Optional. A data.frame describing covariates. This is passed on to DT:data.table() so it can be easily searched

explore_data gives you a simple shiny app to explore data, The app is tailored to the covariate types of the data.frame/data.table that you pass to it.

Each menu item covers a different part of exploratory data analysis with simple visualizations for discussion.

The tabs are automatically updated with variables depending on the variable types.

Value

A Shiny App object that can either be run in the console or on an external shiny server such as shinyapps.io. For building the actual app.R, see build_shiny_app

Examples

library(ggplot2)
data(diamonds)
if(interactive()){
   burro::explore_data(diamonds, outcome_var="cut")
}

#need another example here